python/requirements.txt
flickr-api==0.5
foursquare==20130707
httplib2==0.9
oauth==1.0.1
oauthlib==0.6.3
requests==2.1.0
requests-oauthlib==0.4.0
twython==3.1.2
flickrapi==1.4.4
